clean up compiler warnings

See merge request bitcoinunlimited/BCHUnlimited!2703

+++ src/test/miner_tests.cpp
@@ -354,8 +354,8 @@ void GenerateBlocks(const CChainParams &chainparams,
 
     printf("mempool size : %ld\n", mempool.size());
     printf("mempool mapTx size : %ld\n", mempool.mapTx.size());
-    printf("Avg Block Size %ld Expected Avg Block Size %llu\n", nTotalBlockSize / nBlockCount,
-        nTotalExpectedBlockSize / nBlockCount);
+    printf("Avg Block Size %ld Expected Avg Block Size %lu\n", nTotalBlockSize / nBlockCount,
+        (unsigned long int)(nTotalExpectedBlockSize / nBlockCount));
     printf("Block fill ratio %5.2f\n",
         (double)(nTotalBlockSize / nBlockCount) * 100 / (nTotalExpectedBlockSize / nBlockCount));
     printf("Total mining time: %5.2f\n", (double)nTotalMine / 1000000);
